master: use xrst.git/bin/dev_tools.sh to update the devleopment tools.

+++ bin/group_list.sh
@@ -1,16 +1,25 @@
 #! /usr/bin/env bash
 set -e -u
+# !! EDITS TO THIS FILE ARE LOST DURING UPDATES BY xrst.git/bin/dev_tools.sh !!
 # SPDX-License-Identifier: AGPL-3.0-or-later
 # SPDX-FileCopyrightText: Bradley M. Bell <bradbell@seanet.com>
 # SPDX-FileContributor: 2020-25 Bradley M. Bell
 # ----------------------------------------------------------------------------
 # bin/group_list.sh
 # Returns a space separated list of the group names for this project.
-# The list is extracted from the [root_file] seciton of thexrst.toml
+# The list is extracted from the [root_file] section of the xrst.toml
 # file for this project.
+# If there is no xrst.toml file for this project, the group list is just
+# the default group.
 #
 # sed
 source bin/grep_and_sed.sh
+#
+if [ ! -e xrst.toml ]
+then
+   echo 'default'
+   exit 0
+fi
 #
 # temp.sed
 cat << EOF > temp.sed
